move generated run on example.
tsc, jest and node dist/graphql/index.js all good
