+++ examples/simple/src/ent/address/actions/create_address_action.ts
@@ -1,8 +1,8 @@
 import { AlwaysAllowPrivacyPolicy } from "@snowtop/ent";
 import {
-  CreateAddressActionBase,
-  AddressCreateInput,
-} from "./generated/create_address_action_base";
+  CreateAddressActionBase, 
+  AddressCreateInput, 
+} from "../../generated/address/actions/create_address_action_base";
 
 export { AddressCreateInput };
