+++ examples/simple/src/ent/auth_code/actions/create_auth_code_action.ts
@@ -1,8 +1,8 @@
 import {
-  CreateAuthCodeActionBase,
-  AuthCodeCreateInput,
-} from "./generated/create_auth_code_action_base";
-import { AuthCodeBuilder } from "./generated/auth_code_builder";
+  CreateAuthCodeActionBase, 
+  AuthCodeCreateInput, 
+} from "../../generated/auth_code/actions/create_auth_code_action_base";
+import { AuthCodeBuilder } from "../../generated/auth_code/actions/auth_code_builder";
 
 export { AuthCodeCreateInput };
