+++ examples/simple/src/ent/auth_code/actions/delete_auth_code_action.ts
@@ -1,4 +1,4 @@
-import { DeleteAuthCodeActionBase } from "./generated/delete_auth_code_action_base";
+import { DeleteAuthCodeActionBase } from "../../generated/auth_code/actions/delete_auth_code_action_base";
 
 // we're only writing this once except with --force and packageName provided
 export default class DeleteAuthCodeAction extends DeleteAuthCodeActionBase {}
