+++ examples/simple/src/ent/contact/actions/delete_contact_action.ts
@@ -1,4 +1,4 @@
-import { DeleteContactActionBase } from "./generated/delete_contact_action_base";
+import { DeleteContactActionBase } from "../../generated/contact/actions/delete_contact_action_base";
 
 // we're only writing this once except with --force and packageName provided
 export default class DeleteContactAction extends DeleteContactActionBase {}
