+++ examples/simple/src/ent/contact_email/actions/create_contact_email_action.ts
@@ -12,9 +12,9 @@ import {
 import { AllowIfBuilder } from "@snowtop/ent/action";
 import { Contact } from "../../../ent";
 import {
-  ContactEmailCreateInput,
-  CreateContactEmailActionBase,
-} from "./generated/create_contact_email_action_base";
+  ContactEmailCreateInput, 
+  CreateContactEmailActionBase, 
+} from "../../generated/contact_email/actions/create_contact_email_action_base";
 
 export { ContactEmailCreateInput };
