+++ examples/simple/src/ent/contact_email/actions/delete_contact_email_action.ts
@@ -2,6 +2,6 @@
  * Copyright whaa whaa
  */
 
-import { DeleteContactEmailActionBase } from "./generated/delete_contact_email_action_base";
+import { DeleteContactEmailActionBase } from "../../generated/contact_email/actions/delete_contact_email_action_base";
 
 export default class DeleteContactEmailAction extends DeleteContactEmailActionBase {}
