+++ examples/simple/src/ent/contact_phone_number/actions/delete_contact_phone_number_action.ts
@@ -2,6 +2,6 @@
  * Copyright whaa whaa
  */
 
-import { DeleteContactPhoneNumberActionBase } from "./generated/delete_contact_phone_number_action_base";
+import { DeleteContactPhoneNumberActionBase } from "../../generated/contact_phone_number/actions/delete_contact_phone_number_action_base";
 
 export default class DeleteContactPhoneNumberAction extends DeleteContactPhoneNumberActionBase {}
