+++ examples/simple/src/ent/contact_phone_number/actions/edit_contact_phone_number_action.ts
@@ -3,9 +3,9 @@
  */
 
 import {
-  ContactPhoneNumberEditInput,
-  EditContactPhoneNumberActionBase,
-} from "./generated/edit_contact_phone_number_action_base";
+  ContactPhoneNumberEditInput, 
+  EditContactPhoneNumberActionBase, 
+} from "../../generated/contact_phone_number/actions/edit_contact_phone_number_action_base";
 
 export { ContactPhoneNumberEditInput };
