+++ examples/simple/src/ent/event/actions/edit_event_action.ts
@@ -1,7 +1,7 @@
 import {
-  EditEventActionBase,
-  EventEditInput,
-} from "./generated/edit_event_action_base";
+  EditEventActionBase, 
+  EventEditInput, 
+} from "../../generated/event/actions/edit_event_action_base";
 import { Validator } from "@snowtop/ent/action";
 import { SharedValidators } from "./event_validators";
 import {
@@ -9,7 +9,7 @@ import {
   AlwaysDenyRule,
   PrivacyPolicy,
 } from "@snowtop/ent";
-import { EventBuilder } from "./generated/event_builder";
+import { EventBuilder } from "../../generated/event/actions/event_builder";
 
 export { EventEditInput };
