+++ examples/simple/src/ent/event/actions/edit_event_rsvp_status_action.ts
@@ -1,7 +1,7 @@
 import {
-  EditEventRsvpStatusActionBase,
-  EditEventRsvpStatusInput,
-} from "./generated/edit_event_rsvp_status_action_base";
+  EditEventRsvpStatusActionBase, 
+  EditEventRsvpStatusInput, 
+} from "../../generated/event/actions/edit_event_rsvp_status_action_base";
 
 export { EditEventRsvpStatusInput };
