+++ examples/simple/src/ent/event/actions/event_add_host_action.ts
@@ -1,4 +1,4 @@
-import { EventAddHostActionBase } from "./generated/event_add_host_action_base";
+import { EventAddHostActionBase } from "../../generated/event/actions/event_add_host_action_base";
 
 // we're only writing this once except with --force and packageName provided
 export default class EventAddHostAction extends EventAddHostActionBase {}
