+++ examples/simple/src/ent/event/actions/event_remove_host_action.ts
@@ -1,3 +1,3 @@
-import { EventRemoveHostActionBase } from "./generated/event_remove_host_action_base";
+import { EventRemoveHostActionBase } from "../../generated/event/actions/event_remove_host_action_base";
 
 export default class EventRemoveHostAction extends EventRemoveHostActionBase {}
