+++ examples/simple/src/ent/user/actions/confirm_edit_email_address_action.ts
@@ -2,10 +2,10 @@ import { Trigger, Validator } from "@snowtop/ent/action";
 import DeleteAuthCodeAction from "../../auth_code/actions/delete_auth_code_action";
 import { User } from "../../";
 import {
-  ConfirmEditEmailAddressActionBase,
-  ConfirmEditEmailAddressInput,
-} from "./generated/confirm_edit_email_address_action_base";
-import { UserBuilder } from "./generated/user_builder";
+  ConfirmEditEmailAddressActionBase, 
+  ConfirmEditEmailAddressInput, 
+} from "../../generated/user/actions/confirm_edit_email_address_action_base";
+import { UserBuilder } from "../../generated/user/actions/user_builder";
 
 export { ConfirmEditEmailAddressInput };
