+++ examples/simple/src/ent/user/actions/delete_user_action.ts
@@ -1,4 +1,4 @@
-import { DeleteUserActionBase } from "./generated/delete_user_action_base";
+import { DeleteUserActionBase } from "../../generated/user/actions/delete_user_action_base";
 import { EditUserPrivacy } from "./edit_user_privacy";
 
 // we're only writing this once except with --force and packageName provided
