+++ examples/simple/src/ent/user/actions/edit_phone_number_action.ts
@@ -1,10 +1,10 @@
 import { FakeComms, Mode } from "@snowtop/ent/testutils/fake_comms";
 import CreateAuthCodeAction from "../../auth_code/actions/create_auth_code_action";
 import {
-  EditPhoneNumberActionBase,
-  EditPhoneNumberInput,
-} from "./generated/edit_phone_number_action_base";
-import { UserBuilder } from "./generated/user_builder";
+  EditPhoneNumberActionBase, 
+  EditPhoneNumberInput, 
+} from "../../generated/user/actions/edit_phone_number_action_base";
+import { UserBuilder } from "../../generated/user/actions/user_builder";
 import { User } from "../..";
 import { EditUserPrivacy } from "./edit_user_privacy";
