+++ examples/simple/src/ent/user/actions/edit_user_action.ts
@@ -1,7 +1,7 @@
 import {
-  EditUserActionBase,
-  UserEditInput,
-} from "./generated/edit_user_action_base";
+  EditUserActionBase, 
+  UserEditInput, 
+} from "../../generated/user/actions/edit_user_action_base";
 import { EditUserPrivacy } from "./edit_user_privacy";
 
 export { UserEditInput };
